+++ hydromt/data_catalog/drivers/raster/__init__.py
@@ -1,1 +1,7 @@
 """All drivers that can read RasterDatasets."""
+
+from .raster_dataset_driver import RasterDatasetDriver
+from .raster_xarray_driver import RasterDatasetXarrayDriver
+from .rasterio_driver import RasterioDriver
+
+__all__ = ["RasterDatasetDriver", "RasterDatasetXarrayDriver", "RasterioDriver"]