+++ hydromt/model/components/spatialdatasets.py
@@ -9,9 +9,9 @@ from geopandas import GeoDataFrame
 from pandas import DataFrame
 from xarray import DataArray, Dataset
 
-from hydromt._io.readers import open_ncs
-from hydromt._io.writers import write_nc
 from hydromt._typing.type_def import XArrayDict
+from hydromt.io.readers import open_ncs
+from hydromt.io.writers import write_nc
 from hydromt.model.components.base import ModelComponent
 from hydromt.model.components.spatial import SpatialModelComponent
 from hydromt.model.steps import hydromt_step