+++ hydromt/model/components/vector.py
@@ -12,9 +12,9 @@ from geopandas.testing import assert_geodataframe_equal
 from pyproj import CRS
 from shapely.geometry import box
 
-from hydromt._io.readers import open_ncs
-from hydromt._io.writers import write_nc
 from hydromt.gis.vector import GeoDataset
+from hydromt.io.readers import open_ncs
+from hydromt.io.writers import write_nc
 from hydromt.model.components.base import ModelComponent
 from hydromt.model.components.spatial import SpatialModelComponent
 from hydromt.model.steps import hydromt_step